♻️ [Refactoring]: API型定義を機能別に分割
- api-responses.tsを各エンドポイント別に分割
  - file-responses.ts
  - node-responses.ts
  - component-responses.ts
  - style-responses.ts
  - image-responses.ts
  - comment-responses.ts
  - version-responses.ts
  - team-responses.ts
- api-options.tsを機能別に分割
  - file-options.ts
  - node-options.ts
  - image-options.ts
  - comment-options.ts
- 型定義をapi/ディレクトリに整理
