+++ src/types/api/options/comment-options.ts
@@ -1,0 +1,10 @@
+// コメント関連のAPIオプション型定義
+
+export interface PostCommentOptions {
+  message: string;
+  client_meta: {
+    x: number;
+    y: number;
+  };
+  comment_id?: string;
+}
