+++ src/types/api/options/file-options.ts
@@ -1,0 +1,10 @@
+// ファイル関連のAPIオプション型定義
+
+export interface GetFileOptions {
+  version?: string;
+  ids?: string[];
+  depth?: number;
+  geometry?: 'paths' | 'points';
+  plugin_data?: string;
+  branch_data?: boolean;
+}
