+++ src/types/api/options/node-options.ts
@@ -1,0 +1,9 @@
+// ノード関連のAPIオプション型定義
+
+export interface GetNodesOptions {
+  ids: string[];
+  version?: string;
+  depth?: number;
+  geometry?: 'paths' | 'points';
+  plugin_data?: string;
+}
