+++ src/types/api/responses/comment-responses.ts
@@ -1,0 +1,7 @@
+// コメント関連のAPIレスポンス型定義
+
+import type { Comment } from '../../figma-types.js';
+
+export interface GetCommentsResponse {
+  comments: Comment[];
+}
