+++ src/types/api/responses/image-responses.ts
@@ -1,0 +1,7 @@
+// 画像エクスポート関連のAPIレスポンス型定義
+
+export interface ExportImageResponse {
+  err?: string;
+  images: Record<string, string>;
+  status?: number;
+}
