+++ src/types/api/responses/team-responses.ts
@@ -1,0 +1,17 @@
+// チーム・プロジェクト関連のAPIレスポンス型定義
+
+export interface GetTeamProjectsResponse {
+  projects: Array<{
+    id: string;
+    name: string;
+  }>;
+}
+
+export interface GetProjectFilesResponse {
+  files: Array<{
+    key: string;
+    name: string;
+    thumbnail_url: string;
+    last_modified: string;
+  }>;
+}
