+++ src/types/api/responses/version-responses.ts
@@ -1,0 +1,7 @@
+// バージョン関連のAPIレスポンス型定義
+
+import type { Version } from '../../figma-types.js';
+
+export interface GetVersionsResponse {
+  versions: Version[];
+}
