+++ src/types/index.ts
@@ -1,6 +1,6 @@
 // 型定義の再エクスポート
 
 export * from './figma-types.js';
-export * from './api-responses.js';
-export * from './api-options.js';
+export * from './api/responses/index.js';
+export * from './api/options/index.js';
 export * from './errors.js';
